pyproject.toml
[project]
name = "natis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "amqp==5.3.1",
    "asgiref==3.8.1",
    "billiard==4.2.1",
    "cachetools==5.5.0",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "crispy-tailwind==1.0.3",
    "django==5.1.4",
    "django-allauth==65.3.0",
    "django-appconf==1.0.6",
    "django-browser-reload==1.17.0",
    "django-cotton==1.5.1",
    "django-crispy-forms==2.3",
    "django-environ==0.11.2",
    "django-extensions==3.2.3",
    "django-simple-history==3.7.0",
    "django-storages==1.14.4",
    "djhtml==3.0.7",
    "google-api-core==2.24.0",
    "google-auth==2.37.0",
    "google-cloud-core==2.4.1",
    "google-cloud-storage==2.19.0",
    "google-crc32c==1.6.0",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.66.0",
    "gunicorn==23.0.0",
    "idna==3.10",
    "kombu==5.4.2",
    "packaging==24.2",
    "pillow==11.0.0",
    "prompt-toolkit==3.0.48",
    "proto-plus==1.25.0",
    "protobuf==5.29.3",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "python-dateutil==2.9.0.post0",
    "rcssmin==1.1.2",
    "redis==5.2.1",
    "requests==2.32.3",
    "rjsmin==1.2.2",
    "rsa==4.9",
    "six==1.17.0",
    "sqlparse==0.5.3",
    "tzdata==2024.2",
    "urllib3==2.3.0",
    "vine==5.1.0",
    "wcwidth==0.2.13",
    "whitenoise==6.8.2",
]

[dependency-groups]
dev = [
    "djhtml>=3.0.7",
]
